+++ pkg/cache/utils.go
@@ -1,0 +1,30 @@
+package cache
+
+import (
+	"strconv"
+
+	"k8s.io/klog/v2"
+)
+
+func getPodMetricPort(pod *Pod) int {
+	if pod == nil || pod.Labels == nil {
+		return defaultMetricPort
+	}
+	if v, ok := pod.Labels[MetricPortLabel]; ok && v != "" {
+		if p, err := strconv.Atoi(v); err == nil {
+			return p
+		} else {
+			klog.Warningf("Invalid value for label %s on pod %s/%s: %q. Using default port %d.", MetricPortLabel, pod.Namespace, pod.Name, v, defaultMetricPort)
+		}
+	}
+	return defaultMetricPort
+}
+
+func getPodLabel(pod *Pod, labelName string, defaultValue string) string {
+	labelTarget, ok := pod.Labels[labelName]
+	if !ok {
+		klog.V(4).Infof("No label %v name for pod %v, default to %v", labelName, pod.Name, defaultEngineLabelValue)
+		return defaultValue
+	}
+	return labelTarget
+}
