+++ pkg/plugins/gateway/algorithms/least_busy_time.go
@@ -53,13 +53,13 @@ func (r leastBusyTimeRouter) Route(ctx *types.RoutingContext, readyPodList types
 	minBusyTimeRatio := math.MaxFloat64 // <= 1 in general
 
 	for _, pod := range readyPodList.All() {
-		busyTimeRatio, err := r.cache.GetMetricValueByPodModel(pod.Name, pod.Namespace, ctx.Model, metrics.GPUBusyTimeRatio) // todo: replace mock
+		busyTimeRatio, err := r.cache.GetMetricValueByPod(pod.Name, pod.Namespace, metrics.GPUBusyTimeRatio) // todo: replace mock
 		if err != nil {
 			klog.Error(err)
 			continue
 		}
 		busyTimeRatioValue := busyTimeRatio.GetSimpleValue()
-		klog.Infof("pod: %v, podIP: %v, GPU busy time ratio: %v", pod.Name, pod.Status.PodIP, busyTimeRatioValue)
+		klog.V(4).Infof("pod: %v, podIP: %v, GPU busy time ratio: %v", pod.Name, pod.Status.PodIP, busyTimeRatioValue)
 
 		if busyTimeRatioValue < minBusyTimeRatio {
 			minBusyTimeRatio = busyTimeRatioValue
@@ -74,9 +74,6 @@ func (r leastBusyTimeRouter) Route(ctx *types.RoutingContext, readyPodList types
 		if err != nil {
 			return "", err
 		}
-		klog.Infof("select random pod: %v, podIP: %v", targetPod.Name, targetPod.Status.PodIP)
-	} else {
-		klog.Infof("select target pod: %v, podIP: %v, GPU busy time ratio: %v", targetPod.Name, targetPod.Status.PodIP, minBusyTimeRatio)
 	}
 
 	ctx.SetTargetPod(targetPod)
