+++ pkg/plugins/gateway/algorithms/least_gpu_cache.go
@@ -64,7 +64,7 @@ func (r leastGpuCacheRouter) Route(ctx *types.RoutingContext, readyPodList types
 		}
 		totalCache := gpuCache.GetSimpleValue()
 
-		klog.Infof("pod: %v, podIP: %v, gpuCache: %v",
+		klog.V(4).Infof("pod: %v, podIP: %v, gpuCache: %v",
 			pod.Name, pod.Status.PodIP, gpuCache.GetSimpleValue())
 
 		if totalCache <= minGpuCache {
@@ -80,16 +80,16 @@ func (r leastGpuCacheRouter) Route(ctx *types.RoutingContext, readyPodList types
 		if err != nil {
 			return "", err
 		}
-		klog.Infof("select random targetPod: %s(%s)", targetPod.Name, targetPod.Status.PodIP)
+		klog.V(4).Infof("select random targetPod: %s(%s)", targetPod.Name, targetPod.Status.PodIP)
 	} else {
-		klog.Infof("select targetPod: %s(%s) gpuCache: %v", targetPod.Name, targetPod.Status.PodIP, minGpuCache)
+		klog.V(4).Infof("select targetPod: %s(%s) gpuCache: %v", targetPod.Name, targetPod.Status.PodIP, minGpuCache)
 	}
 
 	if targetPod == nil {
 		return "", fmt.Errorf("no pods to forward request")
 	}
 
-	klog.Infof("targetPod: %s(%s)", targetPod.Name, targetPod.Status.PodIP)
+	klog.V(4).Infof("targetPod: %s(%s)", targetPod.Name, targetPod.Status.PodIP)
 	ctx.SetTargetPod(targetPod)
 	return ctx.TargetAddress(), nil
 }
